Making ContentExtractors first class Baleen objects
Currently, Content extractors are created by the collection reader using only the class name.
This means that the content extract can not use any on the configuration features of UIMA and, in
particualr they can not use the Resources system. Even though the abstract
implementation has the BaleenHistory resource this will always be null.

This commit promotes them to first class objects so they can be included in the Baleen
collection readers using the resource system. This means they can be configured
and have their own resource requirements. It also simpifies the requirements on a
developer to create a new collection reader.

We have moved the configuration to the top level of the yaml eg:

```yaml
contentextractor:
  class:  full.package.if.not.standard.MyContentExctractor.class
  key: "configuration value"
```

we also allow the shorter version where ther is no configuration,

```yaml
contentextractor: PlainTextContentExctractor.class
```

As before the default is the StructureContentExtractor and we still support
the nested declaration of the content extractor for legacy configuration files.

For consistency, in addition, we have allowed the history declaration to follow
the same declaration pattern, either to include the class or not.